fix bugs auth

+++ src/app/interceptors/interceptor.service.ts
@@ -5,14 +5,12 @@ import {
     HttpInterceptor,
     HttpRequest
 } from '@angular/common/http';
-import {Observable, throwError} from 'rxjs';
-import {catchError, map} from 'rxjs/operators';
-import {AuthService} from '../services/auth/auth.service';
-import {Token} from '../models/auth/token';
 import {Router} from '@angular/router';
+import {Observable, throwError} from 'rxjs';
+import {catchError} from 'rxjs/operators';
+import {AuthService} from '../services/auth/auth.service';
+import {Token, Role, System} from '../models/auth/models.index';
 import {Institution} from '../models/app/institution';
-import {Role} from '../models/auth/role';
-import {System} from '../models/auth/system';
 
 @Injectable({
     providedIn: 'root'
@@ -37,22 +35,28 @@ export class InterceptorService implements HttpInterceptor {
                 params = params.append('page', '1');
             }
             if (!req.params.has('per_page')) {
-                params = params.append('page', '10');
+                params = params.append('per_page', '10');
             }
-            if (localStorage.getItem('institution')) {
-                params = params.append('institution',
-                    (JSON.parse(localStorage.getItem('institution')) as Institution).id.toString());
+            if (this.authService.getInstitution()) {
+                params = params.append('institution', this.authService.getInstitution().id.toString());
             }
-            if (localStorage.getItem('role')) {
-                params = params.append('role',
-                    (JSON.parse(localStorage.getItem('role')) as Role).id.toString());
+            if (this.authService.getRole()) {
+                params = params.append('role', this.authService.getRole().id.toString());
+            }
+            if (this.authService.getUri()) {
+                params = params.append('uri', this.authService.getUri());
             }
         }
 
         return next.handle(req.clone({headers, params})).pipe(catchError(error => {
-            if (error.status === 401 || error.status === 423) {
+            if ((error.status === 401 || error.status === 423) && localStorage.getItem('token')) {
                 this.authService.removeLogin();
-                this.router.navigate(['/auth/login']);
+                this.router.navigate(['/auth/access-denied']);
+            }
+
+            if (error.status === 503) {
+                this.authService.removeLogin();
+                this.router.navigate(['/auth/under-maintenance']);
             }
             return throwError(error);
         }));
